requirements.txt
streamlit
langchain-google-genai
langchain-core
python-dotenv
pypdf
SpeechRecognition
edge-tts